+++ dzVents/documentation/history.md
@@ -27,6 +27,7 @@
  - Added compare(time) method to Time object to calculate the difference between them. Returns a table. See documentation.
  - Added round() method to Domoticz object.
  - You can now put your own non-dzVents modules in your scripts folder or write them with the Domoticz GUI editor.
+ - Added humidityStatusValue for humidity devices. This value matches with the values used for setting the humidy status.
 
 [2.2.0]